# docs/swagger.json
{
  "info": {
    "version": "1.0.0",
    "title": "ToDo List API",
    "description": "This is a simple docs for ToDo List API"
  },
  "paths": {
    "/todo": {
      "post": {
        "tags": [
          "ToDo"
        ],
        "summary": "add new ToDo entry",
        "parameters": [
          {
            "in": "body",
            "name": "text",
            "schema": {
              "type": "object",
              "required": [
                "text"
              ],
              "properties": {
                "text": {
                  "type": "string"
                }
              }
            }
          }
        ],
        "description": "",
        "responses": {
          "201": {
            "description": "Created"
          },
          "500": {
            "description": "Internal Server Error"
          }
        }
      },
      "get": {
        "tags": [
          "ToDo"
        ],
        "summary": "Get all ToDo entries",
        "responses": {
          "200": {
            "description": "OK"
          },
          "500": {
            "description": "Internal Server Error"
          },
          "404": {
            "description": "ID Not Found"
          }
        },
        "parameters": []
      }
    },
    "/todo/{ID}": {
      "get": {
        "tags": [
          "ToDo"
        ],
        "summary": "Get the ToDo entry",
        "parameters": [
          {
            "in": "path",
            "name": "ID",
            "description": "ID of the ToDo entry",
            "required": true,
            "type": "integer"
          }
        ],
        "responses": {
          "200": {
            "description": "OK"
          },
          "404": {
            "description": "ID Not Found"
          },
          "500": {
            "description": "Internal Server Error"
          }
        }
      },
      "delete": {
        "tags": [
          "ToDo"
        ],
        "summary": "delete ToDo entry",
        "description": "",
        "parameters": [
          {
            "in": "path",
            "name": "ID",
            "description": "ID of the ToDo entry",
            "required": true,
            "type": "integer"
          }
        ],
        "responses": {
          "204": {
            "description": "No Content"
          },
          "404": {
            "description": "ID Not Found"
          },
          "500": {
            "description": "Internal Server Error"
          }
        }
      },
      "put": {
        "tags": [
          "ToDo"
        ],
        "summary": "update ToDo entry",
        "description": "",
        "parameters": [
          {
            "in": "path",
            "name": "ID",
            "description": "ID of the ToDo entry",
            "required": true,
            "type": "integer"
          },
          {
            "in": "body",
            "name": "text",
            "schema": {
              "type": "object",
              "required": [
                "text"
              ],
              "properties": {
                "text": {
                  "type": "string"
                }
              }
            }
          }
        ],
        "responses": {
          "202": {
            "description": "Accepted"
          },
          "404": {
            "description": "ID Not Found"
          },
          "500": {
            "description": "Internal Server Error"
          }
        }
      }
    }
  },
  "swagger": "2.0",
  "host": "localhost:5000",
  "schemes": [
    "http"
  ]
}